# Make node selection O(n)