--- a/core/dbt/context/common.py
+++ b/core/dbt/context/common.py
@@ -398,8 +398,6 @@ def generate_base(model, model_dict, config, manifest, source_config,
         "exceptions": dbt.exceptions.wrapped_exports(model),
         "execute": provider.execute,
         "flags": dbt.flags,
-        # TODO: Do we have to leave this in?
-        "graph": manifest.to_flat_graph(),
         "log": log,
         "model": model_dict,
         "modules": get_context_modules(),