@drewbanin 2019-07-17
One tiny other thing to note here: since we're always selecting all ephemeral models, dbt won't show a warning if a provided selector doesn't match any models (but > 0 ephemeral models exist in the project):
```
WARNING: Nothing to do. Try checking your model configs and model specification args
```

Can we just change the check in `runnable.py` to warn if there are no non-ephemeral nodes selected?