@beckjake 2019-07-17
@drewbanin I added an early bail instead: if you have selected no nodes, the node selector won't provide add in the ephemeral nodes. That was a lot easier, hope it's ok!