@beckjake 2019-07-17
The windows test failure you'll see in a bit is redshift's concurrent transaction nonsense, I really don't want to re-run those tests.